Implement InnerFromList as separate; Arrange InnerFrom files
