+++ src/collections-flat-array/Collections.FlatArray/FlatArray.T/FlatArray.Factory.Explicit.InnerFromArray.cs
@@ -1,0 +1,8 @@
+﻿namespace System.Collections.Generic;
+
+partial class FlatArray<T>
+{
+    private static FlatArray<T> InnerFromArray(T[] source)
+        =>
+        source.Length > 0 ? new(InnerCloneArray(source), default) : InnerEmptyFlatArray.Value;
+}
