+++ src/collections-flat-array/Collections.FlatArray/FlatArray.T/FlatArray.Factory.Explicit.InnerFromFlatArray.cs
@@ -1,0 +1,8 @@
+﻿namespace System.Collections.Generic;
+
+partial class FlatArray<T>
+{
+    private static FlatArray<T> InnerFromFlatArray(FlatArray<T> source)
+        =>
+        source.items.Length > 0 ? new(InnerCloneArray(source.items), default) : InnerEmptyFlatArray.Value;
+}
