+++ src/collections-flat-array/Collections.FlatArray/FlatArray.T/FlatArray.Factory.Explicit.InnerFromImmutableArray.cs
@@ -1,0 +1,26 @@
+﻿using System.Collections.Immutable;
+
+namespace System.Collections.Generic;
+
+partial class FlatArray<T>
+{
+    private static FlatArray<T> InnerFromImmutableArray(ImmutableArray<T> source)
+    {
+        if (source.IsDefault)
+        {
+            return InnerEmptyFlatArray.Value;
+        }
+
+        var count = source.Length;
+        if (count is not > 0)
+        {
+            return InnerEmptyFlatArray.Value;
+        }
+
+        var array = new T[count];
+        source.CopyTo(array, 0);
+
+        // Clone for the safety purposes
+        return new(InnerCloneArray(array), default);
+    }
+}
