+++ src/collections-flat-array/Collections.FlatArray/FlatArray.T/FlatArray.Factory.Explicit.InnerFromList.cs
@@ -1,0 +1,19 @@
+﻿namespace System.Collections.Generic;
+
+partial class FlatArray<T>
+{
+    private static FlatArray<T> InnerFromList(List<T> source)
+    {
+        var count = source.Count;
+        if (count is not > 0)
+        {
+            return InnerEmptyFlatArray.Value;
+        }
+
+        var array = new T[count];
+        source.CopyTo(array, 0);
+
+        // Clone for the safety purposes
+        return new(InnerCloneArray(array), default);
+    }
+}
